requirements.txt
python-telegram-bot
python-dotenv~=0.19.2
requests~=2.27.1
numerize
